Code review - updated license and refactored towards OOP
