+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2017 Huddle
+Copyright (c) 2018 Huddle
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
