+++ octopose.py
@@ -2,7 +2,7 @@
 
 # MIT License
 #
-# Copyright (c) 2017 Huddle
+# Copyright (c) 2018 Huddle
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -23,120 +23,11 @@
 # SOFTWARE.
 
 import json
-import time
+import remote_deploy
 import argparse
-import shutil
-import os
+import local_deploy
 import sys
 import octo
-import nu
-import util
-
-
-def invoke_deploy(step_path, verbose):
-    """invoke_deploy start a deploy for a given powershell script (*Deploy.ps1)"""
-    if os.path.exists(step_path):
-        print("- {0}".format(step_path))
-        if sys.maxsize > 2**32:
-            args = "powershell.exe {0}".format(step_path)
-        else:
-            args = "c:\\windows\\sysnative\\cmd.exe /c powershell.exe {0}".format(step_path)
-        util.run_subprocess(args, "Running of {0} failed".format(step_path), verbose)
-
-
-def deploy_to_environment(env_id, wait, force, data):
-    """deploy_to_environment will use the manifest to do a remote deploy into another environment"""
-    deployments = {}
-    for key, value in data['Projects'].items():
-        if value is None:
-            continue
-
-        project_name = key
-        proj_id = octo.get_project_id(project_name)
-
-        if 'Version' not in value:
-            latest_release = octo.get_latest_release(proj_id)
-            version = latest_release['Version']
-        else:
-            version = value['Version']
-
-        if version is None:
-            continue
-
-        to_deploy = octo.get_deploy_for_version(proj_id, version)
-        current_deploy = octo.get_deploy_for_env(proj_id, env_id)
-        if version == current_deploy['Version']:
-            last_deploy = octo.get_last_deploy_for_env(proj_id, env_id)
-            last_fail = octo.get_last_failed_deploy_for_env(proj_id, env_id)
-            if (last_fail is not None) and (last_deploy['TaskId'] == last_fail['TaskId']):
-                print("Failed Last Deploy - will try again")
-            elif force:
-                print("Forcing Deploy")
-            else:
-                deployments[project_name] = {'Link': None,
-                                             'Status': 'Already Deployed',
-                                             'Version': version}
-                continue
-
-        if to_deploy:
-            print("{0} - Deploying {1}".format(project_name, version))
-            for package in to_deploy['SelectedPackages']:
-                print("  - {0}".format(package['StepName']))
-            deploy = octo.deploy_release(to_deploy['Id'], env_id)
-            deployments[project_name] = {'Link': deploy['Links']['Task'],
-                                         'Status': 'Queued',
-                                         'Version': version}
-
-    if wait:
-        pending = len(deployments)
-        while pending > 0:
-            for key, value in deployments.items():
-                task_link = value['Link']
-                if task_link is None:
-                    continue
-                task = octo.get_task(task_link)
-                if task['State'] == 'Failed':
-                    deployments[key]['Link'] = None
-                    deployments[key]['Status'] = 'Failed'
-                else:
-                    deployments[key]['Status'] = task['State']
-
-            pending = 0
-            for key, value in deployments.items():
-                status = value['Status']
-                if (status == 'Queued') or (status == 'Executing'):
-                    pending = pending + 1
-                time.sleep(10)
-
-    for key, value in deployments.items():
-        print("{0} - {1} - {2}".format(value['Status'], key,
-                                       value['Version']))
-
-
-def deploy_local(data, verbose):
-    """deploy_local will use the manifest to deploy all packages to the local machine"""
-    staging = os.path.normpath(data['StagingLocation'])
-    if os.path.exists(staging):
-        shutil.rmtree(staging)
-    os.makedirs(staging, mode=0o777)
-
-    for key, value in data['Projects'].items():
-        project_name = key
-        print(project_name)
-        proj_id = octo.get_project_id(project_name)
-        if 'Version' not in value:
-            latest_release = octo.get_latest_release(proj_id)
-            version = latest_release['Version']
-        else:
-            version = value['Version']
-
-        for package in value['Packages']:
-            print("- NuGet - {0}".format(package))
-            nu.get_deployable(package, version, staging, verbose)
-    
-            invoke_deploy("{0}\{1}.{2}\PreDeploy.ps1".format(staging, package, version), verbose)
-            invoke_deploy("{0}\{1}.{2}\Deploy.ps1".format(staging, package, version), verbose)
-            invoke_deploy("{0}\{1}.{2}\PostDeploy.ps1".format(staging, package, version), verbose)
 
 
 if __name__ == "__main__":
@@ -170,7 +61,8 @@ if __name__ == "__main__":
         manifest_string = infile_contents
 
     manifest = json.loads(manifest_string)
+
     if environment != "local":
-        deploy_to_environment(env_id, wait, force, manifest)
+        remote_deploy.deploy_to_environment(env_id, wait, force, manifest)
     else:
-        deploy_local(manifest, verbose)
+        local_deploy.LocalDeploy(verbose).deploy(manifest)
